ex_back/requirements.txt
-i https://pypi.org/simple
asgiref==3.5.2 ; python_version >= '3.7'
django==4.1.2
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
pillow==9.2.0
pytz==2022.4
sqlparse==0.4.3 ; python_version >= '3.5'
